feat/fix: Fix all table unexpected unmount of actions
Issue: When we set a filter in table filter, then we update a specific
data/row and that update did not match the filter, the row itself
including the modal we are editing will disappear/unmounted unexpectedly

This commit brings fixes to all tables that are affected with that bug.
My solution is to take out the state and rendering of modals in those
table rows, instead I mount them directly inside the table itself and
wrap a provider to provide the store that they can consume. So when the
update happen and the data we are focusing disappear, the modal/action
will still be open.

This finally closes #113